Reinstated separate logging helper files
