+++ src/filesystem/ThetisFS.h
@@ -4,9 +4,6 @@
 #include <SD.h>
 #include <SPI.h>
 
-extern char TelemetryLogFilename[32];
-extern char diagnosticLogFilename[32];
-
 bool initSDCard();
 
 // Filesystem Functions
@@ -19,7 +16,3 @@ bool appendFile(fs::FS &fs, const char * path, const char * message);
 bool renameFile(fs::FS &fs, const char * path1, const char * path2);
 bool deleteFile(fs::FS &fs, const char * path);
 void testFileIO(fs::FS &fs, const char * path);
-
-// Logging Functions
-bool initTelemetryLogFile(fs::FS &fs, char * path, char * header);
-
